% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{online_covar}
\alias{online_covar}
\title{online_covar}
\usage{
online_covar(x1, x2)
}
\arguments{
\item{x1}{a numeric vector containing samples of the first random variable}

\item{x2}{a numeric vector containing corresponding samples of the second random variable}
}
\description{
online_covar
}
