% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cor.R
\name{parallel_cor}
\alias{parallel_cor}
\title{Compute correlation function in batches, in parallel}
\usage{
parallel_cor(x, splits = 2, cores = 2, cov_fun = "two_pass_multi_covar")
}
\arguments{
\item{x}{data matrix}

\item{splits}{shows the number of splits of data that we want to combine  (should be of 2^n form)}

\item{cores}{number of cores that will be used for the parallel computation}

\item{cov_fun}{covariance function used on each split}
}
\description{
Compute correlation function in batches, in parallel
}
