% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{two_pass_multi_covar}
\alias{two_pass_multi_covar}
\title{two_pass_multi_covar}
\usage{
two_pass_multi_covar(s)
}
\arguments{
\item{s}{a data matrix whose column covariances are sought}
}
\description{
two_pass_multi_covar
}
